Fixes some guild opcodes for 5.4.0 17399 : - CMSG_GUILD_BANK_LOG_QUERY - CMSG_GUILD_BANK_QUERY_TEXT - CMSG_GUILD_BANK_UPDATE_TAB - CMSG_GUILD_SET_RANK_PERMISSIONS - CMSG_SET_GUILD_BANK_TEXT - SMSG_GUILD_BANK_LOG_QUERY_RESULT - SMSG_GUILD_BANK_QUERY_TEXT_RESULT - CMSG_GUILD_EVENT_LOG_QUERY - SMSG_GUILD_EVENT_LOG_QUERY_RESULT - SMSG_GUILD_XP_GAIN - SMSG_GUILD_ACHIEVEMENT_EARNED - CMSG_GUILD_BANK_WITHDRAW_MONEY - SMSG_GUILD_BANK_MONEY_WITHDRAWN - CMSG_LIST_INVENTORY - SMSG_LIST_INVENTORY - Fix SPELL_EFFECT_UNLOCK_GUILD_VAULT_TAB - CMSG_BUY_ITEM - CMSG_SELL_ITEM - CMSG_BUYBACK_ITEM - SMSG_BUY_ITEM

+++ src/server/game/Scripting/ScriptMgr.cpp
@@ -1388,7 +1388,7 @@ void ScriptMgr::OnGuildDisband(Guild* guild)
     FOREACH_SCRIPT(GuildScript)->OnDisband(guild);
 }
 
-void ScriptMgr::OnGuildMemberWitdrawMoney(Guild* guild, Player* player, uint32 &amount, bool isRepair)
+void ScriptMgr::OnGuildMemberWitdrawMoney(Guild* guild, Player* player, uint64 &amount, bool isRepair)
 {
     FOREACH_SCRIPT(GuildScript)->OnMemberWitdrawMoney(guild, player, amount, isRepair);
 }
